# Optimize SubmissionStepViewSet queries