Using the submission execution state we can both simplify the code and leverage caching of the execution state, which should remove a number of (duplicate) queries.

Checked two cases with django-silk

**`PUT` submission step**

- before this PR: 16 queries, 55ms
- with this PR: 14 queries, 51ms

**`POST` submission step logic check**

- before this PR: 18 queries, 60ms
- with this PR: 14 queries, 49ms

Not terrible, not fantastic - I'm mostly interested in what we'll see in Elastic APM as hints to further optimize.